refactor: 유틸함수 사용하여 리팩토링
