+++ e2e/tests/create-memo.test.ts
@@ -1,28 +1,28 @@
-import { REGEXR } from "../lib/constants";
+import { PATHS } from "@web-memo/shared/constants";
 import { expect, test } from "./fixtures";
+import {
+	fillMemo,
+	findSidePanelPage,
+	LANGUAGE,
+	login,
+	openSidePanel,
+} from "./lib";
 
 test.describe("메모 생성 기능", () => {
-	test.beforeEach(async ({ page, baseURL }) => {
-		await page.goto("/en/memos", {
-			waitUntil: "domcontentloaded",
-		});
-		await page.getByRole("button", { name: "Login with Test" }).click();
-		await page.waitForURL(REGEXR.page.memos);
-
-		await page.goto("https://blog.toss.im/article/toss-team-culture");
-		await page.locator("#OPEN_SIDE_PANEL_BUTTON").click();
-
-		await page.goto(baseURL!);
+	test.beforeEach(async ({ page }) => {
+		await login(page);
+		await openSidePanel(page);
 	});
-	const text = String(new Date());
-	test("사이드 패널에서 메모를 저장하면 메모를 확인할 수 있다.", async ({
+
+	test("사이드 패널에서 메모를 저장하면, 웹사이트에서 메모를 확인할 수 있다.", async ({
 		page,
-		sidePanelPage,
 	}) => {
-		await sidePanelPage.locator("#memo-textarea").fill(text);
-		await page.waitForTimeout(1000);
-		await page.reload();
-		await page.waitForTimeout(1000);
+		const text = String(new Date());
+
+		const sidePanelPage = await findSidePanelPage(page);
+		await fillMemo(sidePanelPage, text);
+
+		await page.goto(`${LANGUAGE}${PATHS.memos}`);
 		await expect(page.getByText(text)).toBeVisible();
 	});
 });
