+++ e2e/tests/fixtures.ts
@@ -1,6 +1,5 @@
-import type { BrowserContext, Page } from "@playwright/test";
+import type { BrowserContext } from "@playwright/test";
 import { test as base, chromium } from "@playwright/test";
-import { PATHS } from "@web-memo/shared/constants";
 import path from "node:path";
 
 process.env.PW_CHROMIUM_ATTACH_TO_OTHER = "1";
@@ -9,8 +8,6 @@ const pathToExtension = path.join(path.resolve(), "..", "dist");
 
 type ExtensionFixture = {
 	context: BrowserContext;
-	sidePanelPage: Page;
-	loginedPage: Page;
 };
 
 export const test = base.extend<ExtensionFixture>({
@@ -26,27 +23,6 @@ export const test = base.extend<ExtensionFixture>({
 		});
 		await use(context);
 		await context.close();
-	},
-	sidePanelPage: async ({ page }, use) => {
-		const sidePanelPage = page
-			.context()
-			.pages()
-			.find(
-				(page) =>
-					page.url() ===
-					"chrome-extension://eaiojpmgklfngpjddhoalgcpkepgkclh/side-panel/index.html",
-			)!;
-		await sidePanelPage.waitForTimeout(1000);
-
-		use(sidePanelPage);
-	},
-	loginedPage: async ({ page }, use) => {
-		await page.goto("/en/login");
-		await page.getByTestId("test-login-button").click();
-		await page.waitForURL(new RegExp(PATHS.memos));
-		await page.waitForTimeout(1000);
-
-		use(page);
 	},
 	baseURL: "http://localhost:3000",
 });
