+++ e2e/tests/lib/constants.ts
@@ -1,0 +1,1 @@
+export const LANGUAGE = "en";
